greybus: tracing: fix host device num_cports
The type of the gb_host_device num_cports field is size_t.  Correct
the num_cports data recorded with a host device event so its type
matches that; fix the format length modifier used for it as well.

+++ drivers/staging/greybus/greybus_trace.h
@@ -303,7 +303,7 @@ DECLARE_EVENT_CLASS(gb_host_device,
 
 	TP_STRUCT__entry(
 		__field(int, bus_id)
-		__field(u8, num_cports)
+		__field(size_t, num_cports)
 		__field(size_t, buffer_size_max)
 	),
 
@@ -313,7 +313,7 @@ DECLARE_EVENT_CLASS(gb_host_device,
 		__entry->buffer_size_max = hd->buffer_size_max;
 	),
 
-	TP_printk("greybus: bus_id=%d num_cports=%hu mtu=%zu",
+	TP_printk("greybus: bus_id=%d num_cports=%zu mtu=%zu",
 		__entry->bus_id, __entry->num_cports,
 		__entry->buffer_size_max)
 );
